requirements.txt
altgraph==0.17
asgiref==3.3.1
attrs==20.3.0
auto-py-to-exe==2.7.11
bcrypt==3.2.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cycler==0.10.0
distro==1.5.0
Django==3.1.4
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==1.0a1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
jwt==1.1.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
openapi-codec==1.3.2
paramiko==2.7.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.26.2
websocket-client==0.57.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.2.0
